+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import re
 from typing import Any, cast
 
 import pytest
@@ -33,6 +34,7 @@ from mex.common.types import (
     YearMonthDay,
 )
 from mex.editor.api.main import api
+from mex.editor.locale_service import LocaleService
 from mex.editor.settings import EditorSettings
 from mex.editor.types import EditorUserDatabase, EditorUserPassword
 
@@ -352,3 +354,14 @@ def load_artificial_extracted_items(
     connector = BackendApiConnector.get()
     connector.ingest(artificial_extracted_items)
     return artificial_extracted_items
+
+
+def build_pagination_regex(current: int, total: int) -> re.Pattern:
+    return re.compile(rf"\w+\s{current}\s\w+\s{total}\s\w+")
+
+
+def build_ui_label_regex(label_id: str) -> re.Pattern:
+    service = LocaleService.get()
+    return re.compile(
+        f"({'|'.join(re.escape(service.get_ui_label(locale.id, label_id)) for locale in service.get_available_locales())})"
+    )
