+++ tests/ingest/test_main.py
@@ -5,7 +5,7 @@ from playwright.sync_api import Page, expect
 
 from mex.common.backend_api.connector import BackendApiConnector
 from mex.editor.ingest.models import ALL_AUX_PROVIDERS, AuxProvider
-from tests.test_utils import build_pagination_regex, build_ui_label_regex
+from tests.conftest import build_pagination_regex, build_ui_label_regex
 
 if TYPE_CHECKING:
     import re
